Implement websockets and basic per-game chat functionality.
Implemented websockets for notifying multiplayer clients
that they need to fetch new data. Message types are UPDATE, CREATE,
JOIN, CLEAR and MESSAGE. The relevant types should be delivered with
in an object with a 'game' key with the value equivalnet to the ID
of a current game.

Chat is setup as an array of subdocuments on the existing game
schema. Re-render prompts are handled by websockets as above.